refactor(selcal): merge audio module and surface invalid-code errors
